Large overhaul of GLM codebase (#319)
* Allow clf to be sklearn object

* Borders are grey even if minmax is larger than range of values

* Add possibility to plot only one hemisphere

* Fixed issue with atlas displaying colors out of bound

* Added pseudo block chance level estimation

* Added kfold interleaved cross-validation

* Middle block blocktrain

* Fixed a bug in middle block training

* Allow for fitting without intercept (sklearn only)

* linear version of GLM code (temporary)

* small changes to defaults

* Beginning to refactor model code

* Slowly getting rid of torch for the moment

* Refactor of GLM

* so many bugs to fix...

* On the road to better tests

* Functional stepwise regression

* A cleansing fire to wash it all away

* Examples to soothe the pain

* Goodbye old friend

* Purgartory is a neverending list of PEP8 errors

* Update release_notes.md
