--- brainbox/modeling/README.md
@@ -1,7 +1,0 @@
-### Welcome to ModelBox!
-
-By Luigi Acerbi and Anne Urai
-
-DataJoint functionality by Shan Shen
-
-International Brain Laboratory, 2019
