+++ release_notes.md
@@ -1,3 +1,8 @@
+## Release Notes 1.11
+### Release Notes 1.11.0
+- brainbox.modeling: New API for using linear and poisson models, doing
+  sequential feature selection.
+
 ## Release Notes 1.10
 ### Release Notes 1.10.0 - 2021-05-04
 - TaskQC: exclude stim_freeze from overall session qc
